creating thel login page, almost completed
